+++ app/src/main/java/codetoon/method/Attack.java
@@ -8,8 +8,7 @@ public class Attack extends MyMethod{
     @Override
     public void action(int i) {
         System.out.println("attack");
-        HostServer.server.updateOpponentTest();
-        GuestServer.server.updateOpponentTest();
+        Server.server.updateOpponentTest();
         Memorys.opponentMemory.get(0).changeColor();
         
     }
