+++ app/src/main/java/codetoon/system/Server.java
@@ -2,8 +2,9 @@ package codetoon.system;
 import java.io.IOException;
 import java.net.Socket;
 import java.io.ObjectOutputStream;
+import java.net.ServerSocket;
 
-abstract class Server implements Runnable {
+public class Server implements Runnable {
     int myPORT = 50000;
     int opponentPORT = 60000;
 
@@ -13,13 +14,66 @@ abstract class Server implements Runnable {
 
     Socket sock;
     Socket returnSock;
+    ServerSocket svSock;
+    ServerSocket svReturnSock;
+
     ObjectOutputStream myOutStream;
     ObjectOutputStream opponentOutStream;
 
     Reception opponentReception;
     Reception returnReception;
 
-    Server() {
+    public static Server server = new Server();
+
+    public void setUpServer() {
+        runServer = true;
+        try {
+            svSock = new ServerSocket(myPORT);
+            svReturnSock = new ServerSocket(opponentPORT);
+            
+            sock = svSock.accept();
+            returnSock = svReturnSock.accept();
+
+            myOutStream = new ObjectOutputStream(sock.getOutputStream());
+            opponentOutStream = new ObjectOutputStream(returnSock.getOutputStream());
+
+            opponentReception = new Reception(sock, false);
+            returnReception = new Reception(returnSock, true);
+            
+            System.out.println("connected");
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        Thread thread = new Thread(this);
+        thread.start();
+    }
+
+    public void connect(String ipAdress) {
+        boolean connect = false;
+        runServer = true;
+        while (connect == false) {
+            try {
+                sock = new Socket(ipAdress, myPORT);
+                returnSock = new Socket(ipAdress, opponentPORT);
+
+                myOutStream = new ObjectOutputStream(sock.getOutputStream());
+                opponentOutStream = new ObjectOutputStream(returnSock.getOutputStream());
+
+                opponentReception = new Reception(sock, false);
+                returnReception = new Reception(returnSock, true);
+                System.out.println("connected");
+                connect = true;
+            } catch (IOException e) {
+                e.printStackTrace();
+                try {
+                    Thread.sleep(5000);
+                } catch (InterruptedException e2) {
+                    e.printStackTrace();
+                }
+            }
+        }
+        Thread thread = new Thread(this);
+        thread.start();
 
     }
 
@@ -97,6 +151,16 @@ abstract class Server implements Runnable {
         returnReception.run();
     }
 
-    abstract public void end();
+    public void end() {
+        runServer = false;
+        try {
+            sock.close();
+            svSock.close();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
+    }
+
 
 }
